Update diagnonsense.md
Small typos mostly, some stylistic formatting stuff, but the big change is from loathesome [sic] to LONESOME. Watching s8ep1 for captions makes me believe it's "lonesome" because of mouth shape/tongue placement. I know that's kind of weird, but I had to watch so many videos in college about that stuff for Russian and linguistics classes. But if you want to hold off on accepting the changes until after Halloween so we can check with Kate, I will totally understand.